action/message: move content under config/action/message
They are no longer needed to be different, because the message action
takes an &mut ResponseSender in its on_*() methods. Before that,
a ResponseSender was wrapped in an Rc<> poiner and that was owned by
the message action.
