+++ src/action.rs
@@ -3,9 +3,7 @@ use state::State;
 use dispatcher::response::ResponseSender;
 use context::base::BaseContext;
 
-pub mod message;
-
-pub use self::message::Alert;
+pub use config::action::message::Alert;
 
 pub trait Action {
     fn on_opened(&self, state: &State, context: &BaseContext, &mut ResponseSender);
@@ -14,6 +12,6 @@ pub trait Action {
 
 pub fn from_config(config: ActionType) -> Box<Action> {
     match config {
-        ActionType::Message(action) => Box::new(self::message::MessageAction::new(action)),
+        ActionType::Message(action) => Box::new(action),
     }
 }
